fix(lore): guard against self-parent loops + SAFE_ID validation on doc tree endpoints
Addresses Devin Review findings on PR #68: parent_doc_id now runs
through the same SAFE_ID pattern as every other primary id in this
file, and both upsertDoc and doc/parent reject parent_doc_id ==
doc_id before touching the DOC_CHILD_OF edge (previously a doc could
be made its own parent, which would loop a tree-walking UI). Also
tightens the MCP sort_order schema to z.number().int() to match the
Java Integer/ArcadeDB INTEGER field it's written to.

+++ mcp-server/src/tools/loreWrite.ts
@@ -903,7 +903,7 @@ export function registerLoreWrite(server: McpServer): void {
         'DeepWiki-style page tree: set this doc\'s parent page in the same call (replaces any existing ' +
         'parent — a doc has at most one). Pass "" (empty string) to detach/move to top level; omit to ' +
         'leave the current parent untouched. For reparenting without touching content, use lore_link_doc_parent instead.'),
-      sort_order: z.number().optional().describe(
+      sort_order: z.number().int().optional().describe(
         'Position among sibling pages under the same parent (used for tree ordering and prev/next navigation).'),
     },
     async (p) => {
